# Conflicts:
#	bsp/boards/gina/uart.c
#	bsp/boards/openmote-cc2538/uart.c
#	bsp/boards/telosb/debugpins.c
#	bsp/boards/telosb/uart.c
#	bsp/boards/wsn430v13b/uart.c
#	bsp/boards/z1/uart.c
#	drivers/common/openserial.c
#	drivers/common/openserial.h
#	openstack/02a-MAClow/IEEE802154E.c
#	openstack/02a-MAClow/IEEE802154_security.c
#	openstack/02b-MAChigh/schedule.c
#	openstack/02b-MAChigh/sf0.c
#	openstack/cross-layers/openqueue.c
#	projects/common/02drv_openserial/02drv_openserial.c
#	projects/common/03oos_sniffer/03oos_sniffer.c
#	projects/openmote-cc2538/03oos_macpong/03oos_macpong.ewp

+++ bsp/boards/uart.h
@@ -48,7 +48,7 @@ void    uart_clearTxInterrupts(void);
 void    uart_setCTS(bool state);
 void    uart_writeByte(uint8_t byteToWrite);
 #ifdef FASTSIM
-void    uart_writeCircularBuffer_FASTSIM(uint8_t* buffer, uint8_t* outputBufIdxR, uint8_t* outputBufIdxW);
+void    uart_writeCircularBuffer_FASTSIM(uint8_t* buffer, uint16_t* outputBufIdxR, uint16_t* outputBufIdxW);
 #endif
 uint8_t uart_readByte(void);
